requirements.txt
streamlit==1.28.1
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
beautifulsoup4==4.12.2
python-dateutil==2.8.2
pandas==2.1.3
lxml==4.9.3
python-dotenv==1.0.0
openai==1.3.7
